+++ adapters/template.pl
@@ -245,6 +245,7 @@ numbervars(Term, From, Next) :-
 '$lgt_default_flag'(duplicated_clauses, silent).
 '$lgt_default_flag'(tail_recursive, silent).
 '$lgt_default_flag'(disjunctions, warning).
+'$lgt_default_flag'(catchall_catch, silent).
 '$lgt_default_flag'(portability, silent).
 '$lgt_default_flag'(redefined_built_ins, silent).
 '$lgt_default_flag'(redefined_operators, warning).